bug: use helm chart namespace when collecting resources
When a ClusterProfile is using configuration drift detection:

1. addon-controller gets list of deployed resources
2. passes this information in a ResourceSummary to drift-detection-manager
3. drift-detection-manager starts watching those resources and when it
detects a configuration drift, drift-detection-manager reports it to
management cluster causing a new reconciliation

With respect to Helm charts, addon-controller gets list of deployed
resources using helm SDK and using manifest.
In certain scenarios, like this [one](https://github.com/projectsveltos/addon-controller/issues/363)
manifest does not contain namespace for namespace resources like deployments.

If namespace is not set for namespace resource, point #3 won't work.

This PR fixes that. When drift-detection-manager gets resources deployed
by addon-controller because of an Helm chart, it adds the helm chart namespace.
This information is later on processed with dynamic.ResourceInterface which
ignores namespace for cluster wide.